Create RouterSimulator_Phase2.py
Developing a network simulator with a user interface where a user can add or remove network components such as routers/ switches and link multiple PC’s. Algorithm developed in Python. This updated code is to simulate a router with multi queuing capability.

The capabilities are M/M/1 and M/D/1 i.e. it can handle random data sizes of packets.